+++ crates/ink/ir/src/ir/item_impl/message.rs
@@ -12,9 +12,33 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+<<<<<<< HEAD
 use super::{ensure_callable_invariants, Callable, CallableKind, InputsIter, Visibility};
 use crate::ir::{self, attrs::SelectorOrWildcard, utils, utils::extract_cfg_attributes};
 use proc_macro2::{Ident, Span, TokenStream};
+=======
+use super::{
+    ensure_callable_invariants,
+    Callable,
+    CallableKind,
+    InputsIter,
+    Visibility,
+};
+use crate::ir::{
+    self,
+    attrs::SelectorOrWildcard,
+    utils,
+    utils::{
+        extract_cfg_attributes,
+        extract_cfg_syn_attributes,
+    },
+};
+use proc_macro2::{
+    Ident,
+    Span,
+    TokenStream,
+};
+>>>>>>> master
 use syn::spanned::Spanned as _;
 
 /// The receiver of an ink! message.
@@ -264,6 +288,11 @@ impl Message {
     /// Returns a list of `cfg` attributes if any.
     pub fn get_cfg_attrs(&self, span: Span) -> Vec<TokenStream> {
         extract_cfg_attributes(self.attrs(), span)
+    }
+
+    /// Returns a list of `cfg` attributes as `syn::Attribute` if any.
+    pub fn get_cfg_syn_attrs(&self) -> Vec<syn::Attribute> {
+        extract_cfg_syn_attributes(self.attrs())
     }
 
     /// Returns the `self` receiver of the ink! message.