+++ crates/ink/tests/ui/contract/pass/cfg-allowed-usage-02.rs
@@ -1,0 +1,27 @@
+#[ink::contract]
+mod contract {
+    #[ink(storage)]
+    pub struct Contract {}
+
+    impl Contract {
+        #[ink(constructor)]
+        pub fn constructor() -> Self {
+            Self {}
+        }
+
+        #[cfg(test)]
+        #[ink(constructor)]
+        pub fn constructor2() -> Self {
+            Self {}
+        }
+
+        #[ink(message)]
+        pub fn message1(&self) {}
+
+        #[ink(message)]
+        #[cfg(test)]
+        pub fn message2(&self) {}
+    }
+}
+
+fn main() {}